backend/requirements.txt
django>=5.0,<6.0
djangorestframework
django-cors-headers
supabase
psycopg2-binary
dj-database-url
gunicorn
python-dotenv
boto3
